move the selector package to common
