+++ sensor/common/selector/selector_test.go
